Moved Firm-related OGFILE3.cpp code to its own file, OGFILE_FIRMS.cpp.
